During the development of #347, @lbenet suggested a way to further reduce allocations in mutating division methods for mixtures (see [this comment](https://github.com/JuliaDiff/TaylorSeries.jl/pull/347#discussion_r1501682516)). After running some local tests, it indeed looks like an option worthwhile pursuing.

On current master:
```julia
julia> using BenchmarkTools

julia> using TaylorSeries
Precompiling TaylorSeries
  1 dependency successfully precompiled in 3 seconds. 3 already precompiled.

julia> dq = set_variables("δq", order=1, numvars=2)
2-element Vector{TaylorN{Float64}}:
  1.0 δq₁ + 𝒪(‖x‖²)
  1.0 δq₂ + 𝒪(‖x‖²)

julia> x = Taylor1([2+2dq[1],-1.1dq[1], 0.7dq[2], 0.5dq[1]-0.45dq[2],0.9dq[1]])
  2.0 + 2.0 δq₁ + 𝒪(‖x‖²) + ( - 1.1 δq₁ + 𝒪(‖x‖²)) t + ( 0.7 δq₂ + 𝒪(‖x‖²)) t² + ( 0.5 δq₁ - 0.45 δq₂ + 𝒪(‖x‖²)) t³ + ( 0.9 δq₁ + 𝒪(‖x‖²)) t⁴ + 𝒪(t⁵)

julia> z = deepcopy(x)
  2.0 + 2.0 δq₁ + 𝒪(‖x‖²) + ( - 1.1 δq₁ + 𝒪(‖x‖²)) t + ( 0.7 δq₂ + 𝒪(‖x‖²)) t² + ( 0.5 δq₁ - 0.45 δq₂ + 𝒪(‖x‖²)) t³ + ( 0.9 δq₁ + 𝒪(‖x‖²)) t⁴ + 𝒪(t⁵)

julia> TaylorSeries.zero!(z)

julia> iszero(z)
true

julia> z[0] = 1/x[0]
 0.5 - 0.5 δq₁ + 𝒪(‖x‖²)

julia> @btime TaylorSeries.div!($z, 1, $x, 1)
  735.992 ns (18 allocations: 1.27 KiB)

julia> @btime TaylorSeries.div!($z, 1, $x, 2)
  1.342 μs (36 allocations: 2.53 KiB)

julia> @btime TaylorSeries.div!($z, 1, $x, 3)
  2.000 μs (54 allocations: 3.80 KiB)

julia> @btime TaylorSeries.div!($z, 1, $x, 4)
  2.657 μs (72 allocations: 5.06 KiB)
```
This PR:
```julia
julia> @btime TaylorSeries.div!($z, 1, $x, 1)
  685.610 ns (18 allocations: 1.27 KiB)

julia> @btime TaylorSeries.div!($z, 1, $x, 2)
  708.029 ns (18 allocations: 1.27 KiB)

julia> @btime TaylorSeries.div!($z, 1, $x, 3)
  724.354 ns (18 allocations: 1.27 KiB)

julia> @btime TaylorSeries.div!($z, 1, $x, 4)
  740.625 ns (18 allocations: 1.27 KiB)
```
~~The remaining 18 allocations are coming from this line:
https://github.com/JuliaDiff/TaylorSeries.jl/blob/bc1735c32546415c1c254a342537c28c361f988b/src/arithmetic.jl#L815~~
EDIT: the afore-mentioned 18 allocations are not there anymore.

UPDATE: after successfully reducing allocation in mutating division and product methods for mixtures, these changes enabled further allocation reduction in sqrt, pow, and other functions. These improvements have also been added to this PR.