@PerezHz 2024-03-04
> @PerezHz I have left three comments on what you have so far in this PR; not sure if they are good ideas or not. In any case, this looks very promising!

Many thanks @lbenet! This is still WIP but even so, many thanks for the comments you left; I'll try to address them in the next few days! I very much agree the changes here can be improved 😄