@PerezHz 2024-03-05
@lbenet I managed to get rid of almost all allocations in `pow!`, except for this line:
https://github.com/PerezHz/TaylorSeries.jl/blob/f95b797a00766396f46f378416129875dbe7528a/src/power.jl#L299
I'm thinking a possible solution could be to simply define a mutating method `power_by_squaring!` for `TaylorN`, waht do you think? Or maybe there's a better way?