@lbenet 2024-03-05
Thanks again, truly, this is looking awesome!!

> @lbenet I managed to get rid of almost all allocations in `pow!`, except for this line: https://github.com/PerezHz/TaylorSeries.jl/blob/f95b797a00766396f46f378416129875dbe7528a/src/power.jl#L299 I'm thinking a possible solution could be to simply define a mutating method `power_by_squaring!` for `TaylorN`, what do you think? Or maybe there's a better way?

While I think it is possible to reduce the number of allocations with respect to the current implementation, the implementation (seems to me) not trivial, because of the recursive nature of the function and the fact that products (and squares) are computed order-by-order. I thus suggest you give it a try, with the possibility to do it in another PR.