@PerezHz 2024-03-06
> While I think it is possible to reduce the number of allocations with respect to the current implementation, the implementation (seems to me) not trivial, because of the recursive nature of the function and the fact that products (and squares) are computed order-by-order. I thus suggest you give it a try, with the possibility to do it in another PR.

This is something I'm still looking at, but perhaps we can leave it as a to-do for the time being. In the end, one allocation per a full expansion is still way better than it was before. I'm still tackling a few remaining cases, but this should be ready for review in the next couple of days...