@PerezHz 2024-03-06
Actually, I think the changes here so far are in a sense self-contained and by themselves achieve good performance improvements. If you agree I'll leave further changes for another PR, so this is now ready for review 😄.