@PerezHz 2024-03-09
> I've left a bunch of comments, mainly in src/functions.jl; they are about inverting (outer) for loops, and if blocks. Other than that, LGTM.

Thank you for spotting this! I've pushed some changes addressing your comments, I'm guessing this time we can bump a patch version, since only the internal API is changing? Tests on TaylorIntegration and NEOs pass with this branch without any changes 👍