@lbenet 2024-03-10
> Thank you for spotting this! I've pushed some changes addressing your comments, I'm guessing this time we can bump a patch version, since only the internal API is changing? Tests on TaylorIntegration and NEOs pass with this branch without any changes 👍

I've left few more comments along the same lines as before. In any case, I do agree with you that we should bump a patch version before merging this PR. It's also very nice that tests pass smoothly in TaylorIntegration.jl and NEOs.jl; have you checked the performance?