@PerezHz 2024-03-10
> have you checked the performance?

So far I have found ~40% improvements in memory usage and performance, sometimes even better!