@lbenet 2024-03-10
> So far I have found ~40% improvements in memory usage and performance, sometimes even better!

Very impressive! Please, let me know when you think this is ready, and I'll go ahead and merge it, and bump the patch version.