@PerezHz 2024-03-10
> Very impressive! Please, let me know when you think this is ready, and I'll go ahead and merge it, and bump the patch version.

Just rebased to current master, and unless CI complains, this is ready to be merged!